Cleanup of threadpool benchmark (#91)
* Moved threadpool benchmarks to a seperate file.

* Formattet the code

* added flamegraph to the threadpool benchmark

* added a constant for the number of requests
